edited hierarchy and makefile, testing makefile on wiliki
