--- sh.exe.stackdump
@@ -1,9 +1,0 @@
-Exception: STATUS_STACK_OVERFLOW at rip=7FFE68DC4B87
-rax=0000000000001250 rbx=00000000FFFFADD0 rcx=0000000000000000
-rdx=0000000180010018 rsi=0000000180271780 rdi=000000018021C490
-r8 =00000000027D3720 r9 =0000000180271780 r10=00000000FFFF9000
-r11=00000000FFE03550 r12=00000000FFFFACB0 r13=000000018021C490
-r14=00000000FFFFAC50 r15=00000000FFFFADD8
-rbp=0000000000000420 rsp=00000000FFFFAAB8
-program=C:\Program Files\Git\usr\bin\sh.exe, pid 10560, thread unknown (0x3304)
-cs=0033 ds=002B es=002B fs=0053 gs=002B ss=002B
